Move platform and platform interface into modules
