+++ src/common.rs
@@ -36,13 +36,14 @@ pub(crate) use crate::{bencode, consts, error, torrent, use_color};
 
 // traits
 pub(crate) use crate::{
-  into_u64::IntoU64, into_usize::IntoUsize, path_ext::PathExt, reckoner::Reckoner,
+  into_u64::IntoU64, into_usize::IntoUsize, path_ext::PathExt,
+  platform_interface::PlatformInterface, reckoner::Reckoner,
 };
 
 // structs and enums
 pub(crate) use crate::{
   env::Env, error::Error, file_info::FileInfo, hasher::Hasher, info::Info, metainfo::Metainfo,
-  mode::Mode, opt::Opt, style::Style, subcommand::Subcommand, torrent::Torrent,
+  mode::Mode, opt::Opt, platform::Platform, style::Style, subcommand::Subcommand, torrent::Torrent,
   use_color::UseColor,
 };
