+++ src/torrent/create.rs
@@ -193,77 +193,6 @@ impl Create {
   }
 }
 
-trait PlatformInterface {
-  fn open(path: &Path) -> Result<(), Error> {
-    let mut command = Self::opener()?;
-    command.push(OsString::from(path));
-
-    let command_string = || {
-      command
-        .iter()
-        .map(|arg| arg.to_string_lossy().into_owned())
-        .collect::<Vec<String>>()
-        .join(",")
-    };
-
-    let status = Command::new(&command[0])
-      .args(&command[1..])
-      .status()
-      .map_err(|source| Error::CommandInvoke {
-        source,
-        command: command_string(),
-      })?;
-
-    if !status.success() {
-      Err(Error::CommandStatus {
-        command: command_string(),
-        status,
-      })
-    } else {
-      Ok(())
-    }
-  }
-
-  fn opener() -> Result<Vec<OsString>, Error>;
-}
-
-struct Platform;
-
-#[cfg(target_os = "windows")]
-impl PlatformInterface for Platform {
-  fn opener() -> Result<Vec<OsString>, Error> {
-    Ok(vec![
-      OsString::from("cmd"),
-      OsString::from("/C"),
-      OsString::from("start"),
-    ])
-  }
-}
-
-#[cfg(target_os = "macos")]
-impl PlatformInterface for Platform {
-  fn opener() -> Result<Vec<OsString>, Error> {
-    Ok(vec![OsString::from("open")])
-  }
-}
-
-#[cfg(not(any(target_os = "windows", target_os = "macos")))]
-impl PlatformInterface for Platform {
-  fn opener() -> Result<(), Error> {
-    const OPENERS: &[&str] = &["xdg-open", "gnome-open", "kde-open"];
-
-    for opener in OPENERS {
-      if let Ok(output) = Command::new(opener).arg("--version").output() {
-        if output.status.success() {
-          return Ok(vec![opener]);
-        }
-      }
-    }
-
-    Err(Error::OpenerMissing { tried: OPENERS })
-  }
-}
-
 #[cfg(test)]
 mod tests {
   use super::*;
